requirements.txt
azure-storage-blob==12.16.0
azure-ai-formrecognizer==3.2.0
scikit-learn==1.3.0
joblib==1.3.2
pytest==7.5.0
